House keeping -- const and types
