+++ src/main/modules/store/bookmarks.ts
@@ -1,6 +1,6 @@
-import { Bookmark } from '@src/types'
 import { Store } from '@main/modules/store/store'
 import { Logger } from '@main/modules/logger'
+import type { Bookmark } from '@src/types'
 
 export class Bookmarks extends Store<{ bookmarks: Bookmark[] }> {
     static instance: Bookmarks
