+++ src/main/modules/store/shortcut.ts
@@ -1,6 +1,6 @@
 import { Store } from '@main/modules/store/store'
 import { SHORTCUTS } from '@main/settings/shortcut'
-import { ShortcutStore } from '@src/types'
+import type { ShortcutStore } from '@src/types'
 
 export class Shortcut extends Store<ShortcutStore> {
     static instance: Shortcut
