+++ src/main/modules/window/abs-window-ipc.ts
@@ -5,15 +5,13 @@ import {
 } from 'electron'
 import { Logger } from '@main/modules/logger'
 
+import type { Scenes, Bookmark, Info, Shortcuts } from '@src/types'
 import {
     Channel,
     RequestHandler,
-    SceneBrowser,
-    type Scenes,
-    type Bookmark,
-    type Info,
-    Shortcuts,
-} from '@src/types'
+    BROWSER,
+    CURRENT_PAGE_INFO,
+} from '@src/constants'
 
 import { Status } from '@main/modules/store/status'
 import { Shortcut } from '@main/modules/store/shortcut'
@@ -21,8 +19,7 @@ import { Anchors } from '@main/modules/store/anchors'
 import { PopupBlocker } from '@src/main/modules/store/popup-blocker'
 import { Bookmarks } from '@main/modules/store/bookmarks'
 
-import { AbsWindowMenu } from './abs-window-menu'
-import { CURRENT_PAGE_INFO } from '@src/constants'
+import { AbsWindowMenu } from '@main/modules/window/abs-window-menu'
 
 /**
  * All starts with here
@@ -101,7 +98,7 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
     ) {
         this.switch(scene)
 
-        if (scene === SceneBrowser.BROWSER && address) {
+        if (this.isBrowser && address) {
             this.title = 'Loading...'
             if (address === 'reload') {
                 this.browser.reload()
@@ -126,7 +123,7 @@ export abstract class AbsWindowIPC extends AbsWindowMenu {
                 return
 
             case RequestHandler.EXECUTE:
-                this.switch(SceneBrowser.BROWSER)
+                this.switch(BROWSER)
                 this.browser.webContents.navigationHistory.goToIndex(index)
                 return
