+++ src/renderer/modules/fragments/button.ts
@@ -1,5 +1,5 @@
-import type { ElementProps } from '@src/types'
 import { Element } from '@home/modules/fragments'
+import type { ElementProps } from '@src/types'
 
 type ButtonTypes = 'submit' | 'reset' | 'button'
 type Props = {
