+++ src/renderer/modules/fragments/form.ts
@@ -1,5 +1,5 @@
-import { ElementProps } from '@src/types'
 import { Element } from '.'
+import type { ElementProps } from '@src/types'
 
 type Props = {
     onSubmit: (ev: HTMLElementEventMap['submit']) => unknown
