+++ src/renderer/modules/fragments/heading.ts
@@ -1,5 +1,5 @@
-import { ElementProps } from '@src/types'
 import { Element } from '.'
+import type { ElementProps } from '@src/types'
 
 export class Heading extends Element<HTMLHeadingElement> {
     constructor(
