+++ src/renderer/modules/fragments/input.ts
@@ -1,5 +1,5 @@
-import type { ElementProps } from '@src/types'
 import { Element } from '.'
+import type { ElementProps } from '@src/types'
 
 type Props = {
     label: string
