+++ src/renderer/modules/fragments/tr-linked.ts
@@ -1,5 +1,5 @@
-import { ElementProps } from '@src/types'
 import { Element } from '.'
+import type { ElementProps } from '@src/types'
 
 export class TrLinked extends Element<HTMLTableRowElement> {
     prev?: TrLinked
