+++ src/renderer/modules/pages/address.ts
@@ -1,6 +1,6 @@
 import { Home } from '@home/modules/pages/home'
 
-import { PageType } from '@src/types'
+import { PageType } from '@src/constants'
 
 export class Address extends Home {
     public readonly page = PageType.ADDRESS
